requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pymongo==4.11.3
bcrypt==4.3.0
python-jose[cryptography]==3.4.0
python-dotenv==1.0.1
twilio==9.5.1
requests==2.32.3
python-multipart==0.0.20
pydantic==2.10.6
pydantic[email]==2.10.6
Jinja2==3.1.6
starlette~=0.46.1
secure==1.0.1
passlib[bcrypt]==1.7.4
celery==5.5.2
celery[redis]==5.5.2
flower==2.0.1
openai==0.28.1
motor
httpx
itsdangerous==2.2.0
schedule==1.2.2